\name{HP-prediction-package}
\alias{HP-prediction-package}
\alias{HP-prediction}
\docType{package}
\title{
\packageTitle{HP-prediction}
}
\description{
\packageDescription{HP-prediction}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HP-prediction}
\packageIndices{HP-prediction}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{HP-prediction}

Maintainer: \packageMaintainer{HP-prediction}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
