% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof.R
\name{lof}
\alias{lof}
\title{A function to left order a matrix based on first occurrence per row}
\usage{
lof(Z, indices = FALSE)
}
\arguments{
\item{Z}{a binary matrix}

\item{indices}{a logical parameter whether to return the column order (`TRUE`) or to return the left-ordered matrix. Defautl is `TRUE`.}
}
\value{
return the left-ordered matrix, or a vector of indexing the order of columns
}
\description{
Given a binary matrix Z with fixed rows, `lof` left orders the matrix sequentially based on first occurrence of each row
}
\examples{
\donttest{
Z <- matrix(rbinom(10*30, 1, 0.1), nrow=10, ncol=30)
Zord = lof(Z)
}

}
