% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammal_supertree}
\alias{mammal_supertree}
\title{Fritz et al. (2009) mammal supertree}
\format{An object of class \code{"phylo"}.}
\source{
\href{https://doi.org/10.1111/j.1461-0248.2009.01307.x}{Ecology Letters}
}
\usage{
data(mammal_supertree)
}
\description{
A phylogeny of 5020 extant mammals, with names harmonized to Wilson & Reeder (2005).
}
\examples{
data(mammal_supertree)
}
\references{
Fritz, S.A., Bininda‐Emonds, O.R.P. and Purvis, A. (2009), 
Geographical variation in predictors of mammalian extinction risk: big is bad, but only in the tropics. 
Ecology Letters, 12: 538-549. \href{https://doi.org/10.1111/j.1461-0248.2009.01307.x}{doi:10.1111/j.1461-0248.2009.01307.x}

Wilson, D.E. & Reeder, D.M. 2005. Mammal Species of the World: A Taxonomic and Geographic Reference. Third Edition. The Johns Hopkins University Press, Baltimore..
}
\keyword{datasets}
