% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_clean.R
\name{network_clean}
\alias{network_clean}
\title{A function to clean the network given multiple pairwise distance matrices or phylo tree}
\usage{
network_clean(
  Z,
  distances,
  model.type = c("full", "distance", "affinity"),
  normalize = TRUE
)
}
\arguments{
\item{Z}{bipartite interaction matrix. Rows should correspond to species in the tree. There should be no empty columns.}

\item{distances}{a single object or a list of phylo objects and symmetric non-negative matrix of pairwise distance between the rows of Z}

\item{model.type}{Indicates the model to use: one of 'full', 'distance', 'affinity'}

\item{normalize}{logical whether to normalize the \code{distances} by maximum distance. (Default and recommended is TRUE).}
}
\value{
Returns a list of the cleaned Z and and list of ordered distances
}
\description{
- excutes \code{network_clean_single} on Z and every passed pairwise distance
- finds intersection species between all distances/phyo trees and keeps only the intersection
- orders the rows of Z according to the cophenetic function
}
