% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_degree.R
\name{plot_degree}
\alias{plot_degree}
\title{A function to plot degree distributions}
\usage{
plot_degree(
  Z,
  Z_est,
  type = "both",
  host.col = "blue",
  parasite.col = "red",
  cex.lab = 1.5,
  cex.axis = 2,
  pt.cex = 1.5,
  legend.cex = 2
)
}
\arguments{
\item{Z}{bipartite interaction matrix}

\item{Z_est}{the posterior bipartite interaction matrix (optional)}

\item{type}{whether to plot 'hosts', 'parasites', or 'both'}

\item{host.col}{Colour to use for hosts (default is 'blue')}

\item{parasite.col}{Colour to use for parasites (default is 'red')}
}
\description{
This function generates a degree distribution plot for the binary interaction matrix 'Z'.
Optionally, you can compare the observed matrix to the posterior matrix by including it as 'Z_est'.
By default the function plots the degree for both rows (hosts) and columns (parasites), but you can modify this with the parameter 'type'.
}
\examples{

# Simluate a Z matrix and plot the degree distributions
\dontrun{
Z <- matrix(rbinom(50*200, 1, 0.01), nrow=50, ncol=200)
Z <- Z[,colSums(Z)>0]
plot_degree(tree)
 }
}
