% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Z.R
\name{plot_Z}
\alias{plot_Z}
\title{A function to plot a bipartite interaction matrix}
\usage{
plot_Z(Z, xlab, ylab, tickMarks = 100, ...)
}
\arguments{
\item{Z}{bipartite interaction matrix}

\item{xlab}{labels for the x axis (default is 'parasites')}

\item{ylab}{labels for the y axis (default is 'hosts')}

\item{...}{Additional parameters that control graphing parameters}
}
\description{
This function generates a raster plot of the binary interaction matrix 'Z' using the 'image' function in the 'graphics' package.
}
\examples{

# Simluate a Z matrix and plot it

Z <- matrix(rbinom(50*200, 1, 0.01), nrow=50, ncol=200)
Z <- Z[,colSums(Z)>0]
plot_Z(tree)

}
