% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurves.R
\name{rocCurves}
\alias{rocCurves}
\title{A function to calculate and plot ROC curves}
\usage{
rocCurves(Z.test, Z.train, P, plot = TRUE, bins = 400, all = FALSE)
}
\arguments{
\item{Z.test}{the bipartite interaction matrix used for the test set}

\item{P}{the posterior probability matrix output by \code{\link{sample_parameter}}}

\item{plot}{TRUE/FALSE to plot the ROC curve.}

\item{bins}{the number of bins that the interval (0,1) is divided into (default is 400)}

\item{all}{TRUE/FALSE to calculate the ROC curve based on the whole dataset, or only the held-out portion}

\item{Z_est}{the estimated bipartite interaction matrix used}
}
\value{
Returns: 
'auc': the maximum AUC value
'threshold': the threshold where P > threshold has the maximum AUC value
'roc': a matrix containing the threthold, FPR, and TPR
}
\description{
Computes the ROC curves as x = false positive rate (FPR) and y = true positive rate (TPR)

FPR = False Positives / (False Positives + True Negatives)
TPR = True Positives / (True Positives + False Negatives)
}
\examples{
 
}
