% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_parameter.R
\name{sample_parameter}
\alias{sample_parameter}
\title{A function to sample the estimated parameters and construct the posterior probability matrix}
\usage{
sample_parameter(model, size = 1000, weights = NULL)
}
\arguments{
\item{model}{model generated by \code{\link{network_est}}}

\item{size}{the number of posterior samples to use for each parameter (default is 1000)}

\item{weights}{indicator of weights to use (optional)}
}
\value{
Returns the posterior probability matrix 'P'
}
\description{
This function samples the affinity and phylogeny scaling parameters from the posterior and constructs the posterior probability matrix 'P'. Dimension names are assigned from matrix `Z`.
}
\examples{
\dontrun{}
 
}
