% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topPairs.R
\name{topPairs}
\alias{topPairs}
\title{A function to list the top predicted interactions from a model}
\usage{
topPairs(P, Z, topX = 10)
}
\arguments{
\item{P}{the estimated posterior probability interaction matrix}

\item{Z}{the observed bipartite interaction matrix}

\item{topX}{number of interactions to return (default is 10)}
}
\value{
Returns a datafame of Host, Parasite, and p(interaction) with length determined by topX (default is 10)
}
\description{
This pulls out the top predicted interactions from a model
}
\examples{
\dontrun{
topPairs(P, Z, topX=15)
}
}
