% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topUndoc.R
\name{topUndoc}
\alias{topUndoc}
\title{A function to list the undocumented interactions with the highest posterior predictions}
\usage{
topUndoc(P, Z, topX = 10)
}
\arguments{
\item{P}{the estimated posterior probability interaction matrix}

\item{Z}{the observed bipartite interaction matrix}

\item{topX}{number of interactions to return (default is 10)}
}
\value{
Returns a datafame of Host, Parasite, and p(interaction) with length determined by topX (default is 10)
}
\description{
This pulls out the top undocumented interactions with the highest posterior prediction.
}
\examples{

topUndoc(P, Z, topX=15)
 
}
