% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammal_supertree}
\alias{mammal_supertree}
\title{A title}
\format{A data frame XXX
\describe{
  \item{ColA}{description about columnA}
}}
\usage{
mammal_supertree
}
\description{
A description
}
\keyword{datasets}
